Only rebuild Spotlight index on first setup

+++ .home/os/setup.sh
@@ -16,37 +16,43 @@ defaults write com.apple.dock wvous-br-corner -int 0
 defaults write com.apple.dock wvous-br-modifier -int 0
 killall Dock
 
-echo "Setting up Spotlight"
-# Change indexing order and disable some search results
-defaults write com.apple.spotlight orderedItems -array \
-	'{"enabled" = 1;"name" = "APPLICATIONS";}' \
-	'{"enabled" = 1;"name" = "SYSTEM_PREFS";}' \
-	'{"enabled" = 1;"name" = "DIRECTORIES";}' \
-	'{"enabled" = 1;"name" = "PDF";}' \
-	'{"enabled" = 1;"name" = "FONTS";}' \
-	'{"enabled" = 0;"name" = "DOCUMENTS";}' \
-	'{"enabled" = 0;"name" = "MESSAGES";}' \
-	'{"enabled" = 0;"name" = "CONTACT";}' \
-	'{"enabled" = 0;"name" = "EVENT_TODO";}' \
-	'{"enabled" = 0;"name" = "IMAGES";}' \
-	'{"enabled" = 0;"name" = "BOOKMARKS";}' \
-	'{"enabled" = 0;"name" = "MUSIC";}' \
-	'{"enabled" = 0;"name" = "MOVIES";}' \
-	'{"enabled" = 0;"name" = "PRESENTATIONS";}' \
-	'{"enabled" = 0;"name" = "SPREADSHEETS";}' \
-	'{"enabled" = 0;"name" = "SOURCE";}' \
-	'{"enabled" = 0;"name" = "MENU_DEFINITION";}' \
-	'{"enabled" = 0;"name" = "MENU_OTHER";}' \
-	'{"enabled" = 0;"name" = "MENU_CONVERSION";}' \
-	'{"enabled" = 0;"name" = "MENU_EXPRESSION";}' \
-	'{"enabled" = 0;"name" = "MENU_WEBSEARCH";}' \
-	'{"enabled" = 0;"name" = "MENU_SPOTLIGHT_SUGGESTIONS";}'
-# Load new settings before rebuilding the index
-killall mds > /dev/null 2>&1
-# Make sure indexing is enabled for the main volume
-sudo mdutil -i on / > /dev/null
-# Rebuild the index from scratch
-sudo mdutil -E / > /dev/null
+if [ ! -f ~/.home/os/.spotlight-reindexed ]; then
+	echo "Setting up Spotlight"
+	# Change indexing order and disable some search results
+	defaults write com.apple.spotlight orderedItems -array \
+		'{"enabled" = 1;"name" = "APPLICATIONS";}' \
+		'{"enabled" = 1;"name" = "SYSTEM_PREFS";}' \
+		'{"enabled" = 1;"name" = "DIRECTORIES";}' \
+		'{"enabled" = 1;"name" = "PDF";}' \
+		'{"enabled" = 1;"name" = "FONTS";}' \
+		'{"enabled" = 0;"name" = "DOCUMENTS";}' \
+		'{"enabled" = 0;"name" = "MESSAGES";}' \
+		'{"enabled" = 0;"name" = "CONTACT";}' \
+		'{"enabled" = 0;"name" = "EVENT_TODO";}' \
+		'{"enabled" = 0;"name" = "IMAGES";}' \
+		'{"enabled" = 0;"name" = "BOOKMARKS";}' \
+		'{"enabled" = 0;"name" = "MUSIC";}' \
+		'{"enabled" = 0;"name" = "MOVIES";}' \
+		'{"enabled" = 0;"name" = "PRESENTATIONS";}' \
+		'{"enabled" = 0;"name" = "SPREADSHEETS";}' \
+		'{"enabled" = 0;"name" = "SOURCE";}' \
+		'{"enabled" = 0;"name" = "MENU_DEFINITION";}' \
+		'{"enabled" = 0;"name" = "MENU_OTHER";}' \
+		'{"enabled" = 0;"name" = "MENU_CONVERSION";}' \
+		'{"enabled" = 0;"name" = "MENU_EXPRESSION";}' \
+		'{"enabled" = 0;"name" = "MENU_WEBSEARCH";}' \
+		'{"enabled" = 0;"name" = "MENU_SPOTLIGHT_SUGGESTIONS";}'
+	# Load new settings before rebuilding the index
+	killall mds > /dev/null 2>&1
+	# Make sure indexing is enabled for the main volume
+	sudo mdutil -i on / > /dev/null
+	# Rebuild the index from scratch
+	sudo mdutil -E / > /dev/null
+	# Only rebuild the index on first setup
+	touch ~/.home/os/.spotlight-reindexed
+else
+	echo "Skipping Spotlight setup"
+fi
 
 echo "Setting up trackpad"
 defaults write com.apple.driver.AppleBluetoothMultitouch.trackpad Clicking -bool true
